Improved render performance
