+++ src/resources/text/index.ts
@@ -1,6 +1,7 @@
 export * from "./cpu-sdf";
-export * from "./font-resource-manager";
-export * from "./glyph-renderer";
 export * from "./font-manager";
 export * from "./font-map";
 export * from "./font-renderer";
+export * from "./font-resource-manager";
+export * from "./font-resource-request";
+export * from "./glyph-renderer";
